+++ tests/functional/test_c.py
@@ -27,37 +27,37 @@ class Test_c(unittest.FunctionalTestCase):
     def test_ce(self):
         self.eq('one |two three', 'ce', 'i_one | three')
         self.eq('one t|wo three', 'ce', 'i_one t| three')
-        self.assertRegister('"', 'wo')
-        self.assertRegister('-', 'wo')
-        self.assertRegister('0', None)
-        self.assertRegister('1', None)
+        self.assertRegister('"wo')
+        self.assertRegister('-wo')
+        self.assertRegisterIsNone('0')
+        self.assertRegisterIsNone('1')
 
     def test_cw(self):
         self.eq('one |two three', 'cw', 'i_one | three')
         self.eq('one t|wo three', 'cw', 'i_one t| three')
-        self.assertRegister('"', 'wo')
-        self.assertRegister('-', 'wo')
-        self.assertRegister('0', None)
-        self.assertRegister('1', None)
+        self.assertRegister('"wo')
+        self.assertRegister('-wo')
+        self.assertRegisterIsNone('0')
+        self.assertRegisterIsNone('1')
 
     def test_c__dollar(self):
         self.eq('one |two three', 'c$', 'i_one |')
         self.eq('one t|wo three', 'c$', 'i_one t|')
-        self.assertRegister('"', 'wo three')
-        self.assertRegister('-', 'wo three')
-        self.assertRegister('0', None)
-        self.assertRegister('1', None)
+        self.assertRegister('"wo three')
+        self.assertRegister('-wo three')
+        self.assertRegisterIsNone('0')
+        self.assertRegisterIsNone('1')
 
     def test_cc(self):
         self.eq('|aaa\nbbb\nccc', 'cc', 'i_|\nbbb\nccc')
         self.eq('aaa\nbb|b\nccc', 'cc', 'i_aaa\n|\nccc')
         self.eq('aaa\nbbb\n|ccc', 'cc', 'i_aaa\nbbb\n|')
-        self.assertRegister('"', 'ccc\n')
-        self.assertRegister('-', None)
-        self.assertRegister('0', None)
-        self.assertRegister('1', 'ccc\n')
-        self.assertRegister('2', 'bbb\n')
-        self.assertRegister('3', 'aaa\n')
+        self.assertRegister('"ccc\n', linewise=True)
+        self.assertRegister('1ccc\n', linewise=True)
+        self.assertRegister('2bbb\n', linewise=True)
+        self.assertRegister('3aaa\n', linewise=True)
+        self.assertRegisterIsNone('-')
+        self.assertRegisterIsNone('0')
 
     def test_cc_should_not_strip_preceding_whitespace(self):
         self.eq('    |one', 'cc', 'i_    |')
@@ -69,28 +69,35 @@ class Test_c(unittest.FunctionalTestCase):
 
     def test_ci__quote(self):
         self.eq('"1|23"', 'ci"', 'i_"|"')
-        self.assertRegister('"', '123')
-        self.assertRegister('-', '123')
-        self.assertRegister('0', None)
-        self.assertRegister('1', None)
+        self.assertRegister('"123')
+        self.assertRegister('-123')
+        self.assertRegisterIsNone('0')
+        self.assertRegisterIsNone('1')
 
     def test_ci__quote_multi_sel(self):
         self.eq('x"1|23"y\ni"ab|c"j\n', 'ci"', 'i_x"|"y\ni"|"j\n')
         self.assertRegister('"', ['123', 'abc'])
         self.assertRegister('-', ['123', 'abc'])
-        self.assertRegister('0', None)
-        self.assertRegister('1', None)
+        self.assertRegisterIsNone('0')
+        self.assertRegisterIsNone('1')
 
     def test_ci__quote__empty_should_not_fill_registers(self):
         self.eq('"|"', 'ci"', 'i_"|"')
-        self.assertRegister('"', None)
-        self.assertRegister('-', None)
-        self.assertRegister('0', None)
-        self.assertRegister('1', None)
+        self.assertRegisterIsNone('"')
+        self.assertRegisterIsNone('-')
+        self.assertRegisterIsNone('0')
+        self.assertRegisterIsNone('1')
 
     def test_ci__quote__multi_sels_empty_should_not_fill_registers(self):
         self.eq('"|"\n"|"', 'ci"', 'i_"|"\n"|"')
-        self.assertRegister('"', None)
-        self.assertRegister('-', None)
-        self.assertRegister('0', None)
-        self.assertRegister('1', None)
+        self.assertRegisterIsNone('"')
+        self.assertRegisterIsNone('-')
+        self.assertRegisterIsNone('0')
+        self.assertRegisterIsNone('1')
+
+    def test_c_visual_line_sets_linewise_register(self):
+        self.eq('x\n|abc\n|y', 'l_c', 'i_x\n|y')
+        self.assertRegister('"abc\n', linewise=True)
+        self.assertRegister('1abc\n', linewise=True)
+        self.assertRegisterIsNone('-')
+        self.assertRegisterIsNone('0')